Split table steps
