+++ frictionless/steps/table/__init__.py
@@ -1,0 +1,15 @@
+from .table_aggregate import table_aggregate
+from .table_attach import table_attach
+from .table_debug import table_debug
+from .table_diff import table_diff
+from .table_intersect import table_intersect
+from .table_join import table_join
+from .table_melt import table_melt
+from .table_merge import table_merge
+from .table_normalize import table_normalize
+from .table_pivot import table_pivot
+from .table_print import table_print
+from .table_recast import table_recast
+from .table_transpose import table_transpose
+from .table_validate import table_validate
+from .table_write import table_write
