+++ frictionless/steps/table/table_normalize.py
@@ -1,0 +1,40 @@
+from ...step import Step
+
+
+# NOTE:
+# We might consider implementing table_preload/cache step
+# Some of the following step use **options - we need to review/fix it
+# Currently, metadata profiles are not fully finished; will require improvements
+# We need to review table_pivot step as it's not fully implemented/tested
+# We need to review table_validate step as it's not fully implemented/tested
+# We need to review table_write step as it's not fully implemented/tested
+# We need to review how we use "target.schema.fields.clear()"
+
+
+class table_normalize(Step):
+    """Normalize table"""
+
+    code = "table-normalize"
+
+    # Transform
+
+    def transform_resource(self, resource):
+        current = resource.to_copy()
+
+        # Data
+        def data():
+            with current:
+                yield current.header.to_list()
+                for row in current.row_stream:
+                    yield row.to_list()
+
+        # Meta
+        resource.data = data
+
+    # Metadata
+
+    metadata_profile = {  # type: ignore
+        "type": "object",
+        "required": [],
+        "properties": {},
+    }
